+++ SwissEphNet/SwissEph.Helpers.cs
@@ -16,7 +16,7 @@ namespace SwissEphNet
         /// <param name="second">Second</param>
         /// <returns>The hour in decimal valeu</returns>
         public static double GetHourValue(int hour, int minute, int second) {
-            return (Double)hour + (minute / 60.0) + (second / 3600.0);
+            return hour + (minute / 60.0) + (second / 3600.0);
         }
 
     }
